Added product search for Corvet table
